requirements.txt
paho-mqtt == 1.5.0
py-dateutil == 2.2
pyserial == 3.4
pyyaml == 5.3
tzlocal == 2.0

